requirements.txt
GitPython==1.0.0
Pillow==2.8.1
